delete colorControl capability, fix formatting in test, change name of test file
